回滚

+++ modules/zh_cn/admin.py
@@ -3,11 +3,10 @@ from graia.ariadne.app import Ariadne
 from graia.ariadne.event.message import GroupMessage
 from graia.ariadne.message.parser.base import DetectPrefix
 from graia.ariadne.model import Group
-from graia.ariadne.util.saya import listen, decorate
+from graia.ariadne.util.saya import listen
 from graia.saya import Channel
 
 import botfunc
-import depen
 
 channel = Channel.current()
 channel.name("管理员")
@@ -16,10 +15,13 @@ channel.author("HanTools")
 
 
 @listen(GroupMessage)
-@decorate(DetectPrefix("上管"))
-@decorate(depen.check_authority_not_black())
-@decorate(depen.check_authority_op())
-async def add_admin(app: Ariadne, group: Group, message: MessageChain = DetectPrefix("上管")):
+async def add_admin(app: Ariadne, group: Group, event: GroupMessage, message: MessageChain = DetectPrefix("上管")):
+    admins = await botfunc.get_all_admin()
+    sb = await botfunc.get_all_sb()
+    if event.sender.id in sb:
+        return
+    if event.sender.id not in admins:
+        return
     try:
         await botfunc.run_sql("INSERT INTO admin(uid) VALUES (%s)", (int(str(message)),))
     except Exception as err:
@@ -29,9 +31,10 @@ async def add_admin(app: Ariadne, group: Group, message: MessageChain = DetectPr
 
 
 @listen(GroupMessage)
-@decorate(DetectPrefix("去管"))
-@decorate(depen.check_authority_bot_op())
-async def del_admin(app: Ariadne, group: Group, message: MessageChain = DetectPrefix("去管")):
+async def del_admin(app: Ariadne, group: Group, event: GroupMessage, message: MessageChain = DetectPrefix("去管")):
+    admins = await botfunc.get_all_admin()
+    if event.sender.id not in admins:
+        return
     try:
         await botfunc.run_sql("DELETE FROM admin WHERE uid = %s", (int(str(message)),))
     except Exception as err:
