+++ modules/zh_cn/blacklist.py
@@ -5,12 +5,11 @@ from graia.ariadne.event.message import GroupMessage
 from graia.ariadne.event.mirai import MemberJoinEvent
 from graia.ariadne.message.element import At
 from graia.ariadne.message.parser.base import DetectPrefix
-from graia.ariadne.util.saya import listen, decorate
+from graia.ariadne.util.saya import listen
 from graia.saya import Channel
 from loguru import logger
 
 import botfunc
-import depen
 
 channel = Channel.current()
 channel.name("黑名单")
@@ -19,9 +18,10 @@ channel.author("HanTools")
 
 
 @listen(GroupMessage)
-@decorate(DetectPrefix("拉黑"))
-@decorate(depen.check_authority_bot_op())
 async def nmsl(app: Ariadne, event: GroupMessage, message: MessageChain = DetectPrefix("拉黑")):
+    admins = await botfunc.get_all_admin()
+    if event.sender.id not in admins:
+        return
     msg = "--- 执行结果 ---\n"
     flag = True
     for i in message[At]:
@@ -71,10 +71,10 @@ async def nmsl(app: Ariadne, event: GroupMessage, message: MessageChain = Detect
 
 
 @listen(MemberJoinEvent)
-@decorate(depen.check_authority_black(False))
 async def kicksb(app: Ariadne, event: MemberJoinEvent):
+    sbs = await botfunc.get_all_sb()
     admins = await botfunc.get_all_admin()
-    if event.inviter.id not in admins:
+    if event.member.id in sbs and event.inviter.id not in admins:
         t = await botfunc.select_fetchone("SELECT uid, op FROM blacklist WHERE uid = %s", (event.member.id,))
         try:
             await app.kick_member(event.member.group)
@@ -86,9 +86,10 @@ async def kicksb(app: Ariadne, event: MemberJoinEvent):
 
 
 @listen(GroupMessage)
-@decorate(DetectPrefix("删黑"))
-@decorate(depen.check_authority_bot_op())
 async def nmms(app: Ariadne, event: GroupMessage, message: MessageChain = DetectPrefix("删黑")):
+    admins = await botfunc.get_all_admin()
+    if event.sender.id not in admins:
+        return
     try:
         await botfunc.run_sql('DELETE FROM blacklist WHERE uid = %s',
                               (int(str(message)),))
